+++ src/components/UI/Records/features/Transfer/Transfer.tsx
@@ -24,6 +24,8 @@ import { FormContainer, SecondaryButtonForm } from '../RecordTemplate/RecordTemp
 import { ShowExpenses } from '../ShowExpenses';
 import { FlexContainer } from '../../../../../styles';
 import { SelectExpenses } from '../SelectExpenses';
+import { useGuestUser } from '../../../../../hooks/useGuestUser/useGuestUser';
+import { EditExpenseProps, EditIncomeProps } from '../../../../../hooks/useRecords/interface';
 
 interface TransferProps {
   action: string;
@@ -37,12 +39,15 @@ dayjs.extend(timezone);
 const Transfer = ({ action, typeOfRecord, edit = false }: TransferProps) => {
   const {
     createTransfer,
+    createTransferLocal,
     editExpense,
     editIncome,
+    editTransferLocal,
     isLoadingCreateTransfer,
     isSuccessCreateTransfer,
   } = useRecords({});
   const { initialAmount, updateAmount, verifyAmountEndsPeriod } = useCurrencyField();
+  const { isGuestUser } = useGuestUser();
 
   const recordToBeEdited = useAppSelector((state) => state.records.recordToBeModified);
   const isIncome = !!recordToBeEdited?.expensesPaid;
@@ -150,19 +155,24 @@ const Transfer = ({ action, typeOfRecord, edit = false }: TransferProps) => {
 
     const previousAmount = recordToBeEdited?.amount ?? 0;
     const userIdRecord = recordToBeEdited?.userId ?? '';
-    resetLocalStorageWithUserOnly();
-    await editExpense({
+
+    const expensePayload: EditExpenseProps = {
       values: newValuesExpense,
       recordId: recordIdExpense,
       amountTouched,
       previousAmount,
       userId: userIdRecord,
       accountId: expenseAccount,
-    });
+    };
+
+    if (!isGuestUser) {
+      resetLocalStorageWithUserOnly();
+      await editExpense(expensePayload);
+    }
 
     const previousExpensesRelated = recordToBeEdited?.expensesPaid ?? [];
     const { oldRecords } = symmetricDifferenceExpensesRelated(previousExpensesRelated, expensesSelected);
-    await editIncome({
+    const incomePayload: EditIncomeProps = {
       values: newValuesIncome,
       recordId: recordIdIncome,
       amountTouched,
@@ -170,7 +180,13 @@ const Transfer = ({ action, typeOfRecord, edit = false }: TransferProps) => {
       previousExpensesRelated: oldRecords,
       userId: userIdRecord,
       accountId: incomeAccount,
-    });
+    };
+
+    if (!isGuestUser) {
+      await editIncome(incomePayload);
+    } else {
+      editTransferLocal({ payloadIncome: incomePayload, payloadExpense: expensePayload });
+    }
   };
 
   const handleCreateTransfer = (values: CreateTransferValues) => {
@@ -178,6 +194,12 @@ const Transfer = ({ action, typeOfRecord, edit = false }: TransferProps) => {
     const newAmount = verifyAmountEndsPeriod(initialAmount.current);
     const newValues = { ...restValues, amount: newAmount };
     const { newValuesIncome, newValuesExpense } = getValuesIncomeAndExpense({ values: newValues, expensesSelected });
+
+    if (isGuestUser) {
+      createTransferLocal({ valuesExpense: newValuesExpense, valuesIncome: newValuesIncome });
+      return;
+    }
+
     createTransfer({ valuesExpense: newValuesExpense, valuesIncome: newValuesIncome });
   };