+++ src/components/templates/Header/interface.ts
@@ -1,0 +1,10 @@
+export interface HeaderProps {
+  isLandingPage?: boolean;
+}
+
+export interface HeaderShadowProps {
+  isLandingPage: boolean;
+}
+export interface GuestUserButtonProps {
+  isLandingPage: boolean;
+}