+++ src/hooks/useAllExpenses/useAllExpenses.tsx
@@ -1,22 +1,66 @@
-import { useMemo } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { GET_EXPENSES } from '../../components/UI/Records/constants';
 import { UseAllExpensesProps } from './interface';
 import { useAppSelector } from '../../redux/hooks';
 import { useGetExpensesQuery } from '../../redux/slices/Records/actions/expenses.api';
+import { Expense } from '../../globalInterface';
+import { useDate } from '../useDate';
+import { getLocalRecords } from './utils';
 
 const useAllExpenses = ({ month, year, accountId }: UseAllExpensesProps) => {
+  const { month: currentMonth, lastMonth } = useDate();
   const userReduxState = useAppSelector((state) => state.user);
+  const isGuestUser: boolean = userReduxState?.userInfo?.user?.firstName === 'Guest';
+  const recordsLocalStorage = useAppSelector((state) => state.records.recordsLocalStorage);
   const bearerToken = userReduxState.userInfo?.bearerToken as string;
   const selectedAccount = useAppSelector((state) => state.accounts.accountSelected);
   const selectedAccountId = accountId ?? selectedAccount?._id;
   const fullRoute = `${GET_EXPENSES}/${selectedAccountId}/${month}/${year}`;
+  const recordsLocalStorageSelectedAccount = recordsLocalStorage?.find((record) => record.account === selectedAccountId);
 
-  const { isFetching, isError, currentData } = useGetExpensesQuery({ route: fullRoute, bearerToken }, { skip: (!bearerToken || !selectedAccountId) });
-  const onlyExpensesIncomes = useMemo(() => currentData?.records.filter((record) => record.typeOfRecord !== 'transfer'), [currentData?.records]);
+  const [localRecords, setLocalRecords] = useState<Expense[]>([]);
+  const [noExpensesFound, setNoExpensesFound] = useState<boolean>(false);
+
+  const toggleNoExpensesFound = () => {
+    setNoExpensesFound((prevState) => !prevState);
+  };
+
+  useEffect(() => {
+    if (isGuestUser && recordsLocalStorage) {
+      const fetchedLocalRecords = getLocalRecords({
+        month,
+        lastMonth,
+        currentMonth,
+        year,
+        recordsLocalStorageSelectedAccount,
+        toggleNoExpensesFound,
+        noExpensesFound,
+      });
+      setLocalRecords(fetchedLocalRecords);
+    }
+  }, [currentMonth, isGuestUser, lastMonth, month, noExpensesFound, recordsLocalStorage, recordsLocalStorageSelectedAccount, year]);
+
+  const { isFetching, isError, currentData } = useGetExpensesQuery(
+    { route: fullRoute, bearerToken },
+    { skip: (!bearerToken || !selectedAccountId || isGuestUser) },
+  );
+
+  useEffect(() => {
+    if (currentData?.message === 'No expenses found.') {
+      toggleNoExpensesFound();
+    }
+  }, [currentData?.message]);
+
+  const onlyExpensesIncomes = useMemo(
+    () => (currentData?.records ?? []).filter((record) => record.typeOfRecord === 'expense'),
+    [currentData?.records],
+  );
+
+  const recordsToShow = isGuestUser ? localRecords : onlyExpensesIncomes;
 
   return {
-    expenses: onlyExpensesIncomes ?? [],
-    noExpensesFound: currentData?.message === 'No expenses found.',
+    expenses: recordsToShow,
+    noExpensesFound,
     isError,
     loading: isFetching,
   };