+++ src/hooks/useGuestUser/index.ts
@@ -1,0 +1,1 @@
+export { useGuestUser } from './useGuestUser';