+++ src/pages/RoutesConstants.ts
@@ -1,6 +1,7 @@
-export const LOGIN_ROUTE = '/';
+export const LOGIN_ROUTE = '/login';
 export const FORGOT_PASSWORD_ROUTE = '/forgot-password';
 export const REGISTER_ROUTE = '/register';
 export const DASHBOARD_ROUTE = '/dashboard';
+export const LANDING_ROUTE = '/';
 export const CREATE_RECORD_ROUTE = '/create-record';
 export const EDIT_RECORD_ROUTE = '/edit-record';