+++ src/utils/LocalStorage/addToLocalStorage.ts
@@ -1,12 +1,20 @@
 import { getLocalStorageInfo } from './getLocalStorageInfo';
 import { saveInfoToLocalStorage } from './saveInfoToLocalStorage';
 
-export function addToLocalStorage(newInfo: object) {
+export function addToLocalStorage({ newInfo, prop }: { newInfo: object, prop?: string }) {
   const localStorageInfo = getLocalStorageInfo();
-  const newLocalStorage = {
-    ...localStorageInfo,
-    ...newInfo,
-  };
+  let newLocalStorage;
+  if (prop) {
+    newLocalStorage = {
+      ...localStorageInfo,
+      [prop]: newInfo,
+    };
+  } else {
+    newLocalStorage = {
+      ...localStorageInfo,
+      ...newInfo,
+    };
+  }
   saveInfoToLocalStorage(newLocalStorage);
   return newLocalStorage;
 }