+++ src/utils/LocalStorage/interface.ts
@@ -1,10 +1,27 @@
-import { AnyRecord, User } from '../../globalInterface';
+import {
+  Account, AnyRecord, RecordRedux, User,
+} from '../../globalInterface';
 
-export interface CountOnMeLocalStorage {
+export interface RecordAgeStatus {
+  currentMonth: RecordRedux[];
+  lastMonth: RecordRedux[];
+  olderRecords: RecordRedux[];
+}
+
+export type RecordAgeStatusKey = keyof RecordAgeStatus;
+
+export interface RecordsLocalStorage {
+  account: string;
+  records: RecordAgeStatus;
+}
+
+export interface BudgetMasterLocalStorage {
   bearerToken: string
   accessToken: string
   user: User;
   recordToBeEdited: AnyRecord;
+  accounts?: Account[];
+  records?: RecordsLocalStorage[];
 }
 
 export interface JWT {