+++ src/utils/LocalStorage/resetLocalStorageWithUserOnly.ts
@@ -1,9 +1,9 @@
 import { getLocalStorageInfo } from './getLocalStorageInfo';
-import { CountOnMeLocalStorage } from './interface';
+import { BudgetMasterLocalStorage } from './interface';
 import { saveInfoToLocalStorage } from './saveInfoToLocalStorage';
 
 export function resetLocalStorageWithUserOnly() {
-  const localStorageInfo: CountOnMeLocalStorage = getLocalStorageInfo();
+  const localStorageInfo: BudgetMasterLocalStorage = getLocalStorageInfo();
   const { user, bearerToken, accessToken } = localStorageInfo;
   saveInfoToLocalStorage({ user, bearerToken, accessToken });
 }